+++ middlepj/.classpath
@@ -6,9 +6,47 @@
 			<attribute name="module" value="true"/>
 		</attributes>
 	</classpathentry>
-
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/cos.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/gson-2.9.0.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/ojdbc6_g.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/taglibs-standard-compat-1.2.5.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/taglibs-standard-impl-1.2.5.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/taglibs-standard-jstlel-1.2.5.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/admin/git/middleproject/middlepj/src/main/webapp/WEB-INF/lib/taglibs-standard-spec-1.2.5.jar">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="con" path="org.eclipse.jst.server.core.container/org.eclipse.jst.server.tomcat.runtimeTarget/Apache Tomcat v9.0">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.web.container"/>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.module.container"/>
-	<classpathentry kind="con" path="org.eclipse.jst.server.core.container/org.eclipse.jst.server.tomcat.runtimeTarget/apache-tomcat-9.0.55"/>
 	<classpathentry kind="output" path="build/classes"/>
 </classpath>
