+++ middlepj/src/main/java/middle/dao/qnaDAO.java
@@ -1,0 +1,78 @@
+package middle.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+import middle.vo.qnaVO;
+
+public class qnaDAO extends DAO{
+	
+	Connection conn;
+	ResultSet rs;
+	 PreparedStatement psmt; 
+	 
+	
+//	 public void write(String qnaTitle, String qnaContent, String id, qnaVO qna) {
+//		 conn = getConnect();
+//		 String sql = "insert into qna(qna_date,qna_no,qna_pw,qna_category,product_code,qna_writer,qna_title,qna_content,id) values(?,?,?,?,?,?,?,?,?)";
+//		 try {
+//			 psmt.setString(1, qna.getQnaDate());
+//			 psmt.setInt(2, qna.getQnaNo());
+//			 psmt.setString(3, qna.getQnaPw());
+//			 psmt.setString(4, qna.getQnaCategory());
+//			 psmt.setString(5, qna.getProductCode());
+//			 psmt.setString(6, qna.getQnaWrite());
+//			 psmt.setString(7, qnaTitle);
+//			 psmt.setString(8, qnaContent);
+//			 psmt.setString(9, id);
+//			 int r = psmt.executeUpdate();
+//			 System.out.println(r + "건 입력되었습니다");
+//		 }catch(SQLException e) {
+//			 e.printStackTrace();
+//		 }finally {
+//			 disconnect();
+//		 }
+//	 }
+//	 
+
+	public List<qnaVO> getList() {
+		conn = getConnect();
+		getConnect();
+		List<qnaVO> list = new ArrayList<qnaVO> ();
+		
+		try {
+			psmt = conn.prepareStatement("select*from qna order by qna_no desc");
+			rs = psmt.executeQuery();
+			while(rs.next()) {
+				qnaVO qnavo = new qnaVO();
+				qnavo.setQnaCategory(rs.getString("qna_category"));
+				qnavo.setQnaNo(rs.getInt("qna_no"));
+				qnavo.setProductCode(rs.getString("product_code"));
+				qnavo.setId(rs.getString("id"));
+				qnavo.setQnaTitle(rs.getString("qna_title"));
+				qnavo.setQnaContent(rs.getString("qna_content"));
+				qnavo.setQnaWrite(rs.getString("qna_writer"));
+				qnavo.setQnaDate(rs.getString("qan_date"));
+				qnavo.setQnaPw(rs.getString("qna_pw"));
+				list.add(qnavo);
+				
+			}
+		}catch (SQLException e) {
+			e.printStackTrace();
+		}finally {
+			disconnect();
+		}
+		return list;
+	}
+	
+	public void writeQna(qnaVO qna) {
+		conn = getConnect();
+		getConnect();
+		String sql = "";
+	}
+	
+}
