+++ middlepj/src/main/java/middle/vo/qnaVO.java
@@ -1,4 +1,4 @@
-package qna;
+package middle.vo;
 
 public class qnaVO {
