--- middlepj/src/main/java/qna/Control.java
@@ -1,11 +1,0 @@
-package qna;
-
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public interface Control {
-	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
-}
