+++ middlepj/src/main/webapp/WEB-INF/web.xml
@@ -20,6 +20,6 @@
 	</servlet-mapping>
 
 	<welcome-file-list>
-		<welcome-file>index.jsp</welcome-file>
+		<welcome-file>view/main.jsp</welcome-file>
 	</welcome-file-list>
 </web-app>
